[Refactor] @JasungLee 리뷰 반영
