+++ core/designsystem/src/main/kotlin/team/ppac/designsystem/util/extension/BoxShadow.kt
@@ -1,12 +1,7 @@
 package team.ppac.designsystem.util.extension
 
-import android.annotation.SuppressLint
 import android.graphics.BlurMaskFilter
-import androidx.compose.foundation.clickable
-import androidx.compose.foundation.interaction.MutableInteractionSource
-import androidx.compose.runtime.remember
 import androidx.compose.ui.Modifier
-import androidx.compose.ui.composed
 import androidx.compose.ui.draw.clip
 import androidx.compose.ui.draw.drawWithCache
 import androidx.compose.ui.geometry.Size
@@ -24,21 +19,6 @@ import androidx.compose.ui.unit.Dp
 import androidx.compose.ui.unit.DpOffset
 import androidx.compose.ui.unit.dp
 import androidx.compose.ui.unit.isSpecified
-
-@SuppressLint("ModifierFactoryUnreferencedReceiver")
-fun Modifier.noRippleClickable(
-    enabled: Boolean = true,
-    onClick: () -> Unit,
-    debounceMillis: Long = 300L,
-): Modifier = composed {
-    val multipleEventsCutter = remember { MultipleEventsCutter(debounceMillis) }
-    clickable(
-        indication = null,
-        enabled = enabled,
-        interactionSource = remember { MutableInteractionSource() },
-        onClick = { multipleEventsCutter.processEvent(onClick) },
-    )
-}
 
 fun Modifier.boxShadow(
     color: Color,
